refactor: share function to remove repo owner from full repo name
